Finalizando Busca de Pedidos

+++ lib/models/section.dart
@@ -75,7 +75,7 @@ class Section extends ChangeNotifier {
     }
 
     for(final original in originalItems){
-      if(!items.contains(original)){
+      if(!items.contains(original) && (original.image as String).contains('firebase')){
         try {
           final ref = await storage.getReferenceFromUrl(
               original.image as String
@@ -97,12 +97,14 @@ class Section extends ChangeNotifier {
   Future<void>delete() async {
     await firestoreRef.delete();
     for(final item in items){
+      if((item.image as String).contains('firebase')){
       try{
         final ref = await storage.getReferenceFromUrl(
-            item.image as String
+          item.image as String
         );
-        await ref.delete();
-      } catch(e){}
+          await ref.delete();
+        } catch(e){}
+      }
     }
   }
